refactor: move consents to own store submodule
