+++ src/views/PrivacyPolicy.vue
@@ -515,7 +515,7 @@
           <v-layout column align-center>
             <v-form>
               <v-flex
-                v-for="option in this.$store.state.session.cookieOptions"
+                v-for="option in this.$store.state.consents.cookieOptions"
                 :key="option.category"
               >
                 <v-checkbox
@@ -546,7 +546,7 @@
 
 <script lang="ts">
 import Vue from "vue";
-import { Consent, CookieOption } from "@/store/modules/session";
+import { Consent, CookieOption } from "@/store/modules/consents";
 
 export default Vue.extend({
   name: "PrivacyPolicy",
@@ -561,16 +561,17 @@ export default Vue.extend({
     isSettingsChangeSuccess: false
   }),
   created() {
-    this.$store.state.session.cookieOptions.forEach((option: CookieOption) => {
+    // Load up checkbox values from default and user's consent values
+    this.$store.state.consents.cookieOptions.forEach((option: CookieOption) => {
       this.cookies[option.category] = option.default;
     });
-    this.$store.state.session.consentsPromise.then(() => {
-      this.$store.state.session.consents
+    this.$store.state.consents.consentsPromise.then(() => {
+      this.$store.state.consents.consents
         .filter(({ category }: Consent) => category !== "strictly_necessary")
         .forEach((consent: Consent) => {
           if (this.cookies.hasOwnProperty(consent.category)) {
             this.cookies[consent.category] = this.$store.getters[
-              "session/isConsentAccepted"
+              "consents/isConsentAccepted"
             ](consent);
           }
         });
@@ -581,9 +582,9 @@ export default Vue.extend({
       this.showForm = this.showForm ? false : true;
     },
     acceptPrivacyChanges() {
-      this.$store.state.session.cookieOptions.forEach(
+      this.$store.state.consents.cookieOptions.forEach(
         ({ category, message }: CookieOption) => {
-          this.$store.dispatch("session/addCookieConsent", {
+          this.$store.dispatch("consents/addCookieConsent", {
             category: category,
             message: message,
             status: this.cookies[category] ? "accepted" : "rejected",
